Made Jetpack.fallDamageEvent accept doubles, for 1.6 and above

+++ src/com/github/jselby/easyjetpack/Jetpack.java
@@ -165,7 +165,7 @@ public class Jetpack {
 	}
 
 	// Fall damage event
-	public static boolean fallDamageEvent(Player player, int amount) {
+	public static boolean fallDamageEvent(Player player, double amount) {
 		ConfigHandler.checkConfig();
 		boolean cancel = false;
 		if (player.hasPermission("easyjetpack.softlanding")
